processing_service/requirements.txt
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
grpcio==1.74.0
grpcio-tools==1.74.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
